+++ pkg/analyzer/typechecker/tests/function_call_resolution_test.go
@@ -31,8 +31,8 @@ func TestCall_UndefinedFunction_ResultUsedInBinaryExpr_NoError_Cascade(t *testin
 }
 
 func TestCall_UndefinedFunction_AssignedToAnnotatedVar_NoError_Cascade(t *testing.T) {
-	// checkVarDecl bails on nil inferredType, so no "cannot assign … to int" cascade.
-	res := parseCollectAndCheck(t, `let y: int = foo()`, false)
+	// checkVarDecl bails on nil inferredType, so no "cannot assign … to i64" cascade.
+	res := parseCollectAndCheck(t, `let y: i64 = foo()`, false)
 	assertErrorsAre(t, res, `undefined function "foo"`)
 }
 
@@ -62,7 +62,7 @@ func TestCall_UndefinedFunction_NestedAsArgument_OnlyOneError(t *testing.T) {
 	// The outer call (double) resolves fine; the inner call (foo) is undefined.
 	// The argument type check is skipped when argType is nil, so only one error.
 	res := parseCollectAndCheck(t, `
-		let double = (n: int) -> int => n * 2
+		let double = (n: i64) -> i64 => n * 2
 		double(foo())
 	`, false)
 	assertErrorsAre(t, res, `undefined function "foo"`)
@@ -71,7 +71,7 @@ func TestCall_UndefinedFunction_NestedAsArgument_OnlyOneError(t *testing.T) {
 func TestCall_UndefinedFunction_Typo_DefinedNameNotFound(t *testing.T) {
 	// "add" is defined but "adds" is not — a one-character typo must not resolve.
 	res := parseCollectAndCheck(t, `
-		let add = (a: int, b: int) -> int => a + b
+		let add = (a: i64, b: i64) -> i64 => a + b
 		adds(1, 2)
 	`, false)
 	assertErrorsAre(t, res, `undefined function "adds"`)
@@ -84,7 +84,7 @@ func TestCall_NonCallable_IntVariable(t *testing.T) {
 		let n = 42
 		n()
 	`, false)
-	assertErrorsAre(t, res, `identifier "n" is not callable (type int)`)
+	assertErrorsAre(t, res, `identifier "n" is not callable (type i64)`)
 }
 
 func TestCall_NonCallable_StringVariable(t *testing.T) {
@@ -109,7 +109,7 @@ func TestCall_NonCallable_WithArguments(t *testing.T) {
 		let n = 42
 		n(1, 2)
 	`, false)
-	assertErrorsAre(t, res, `identifier "n" is not callable (type int)`)
+	assertErrorsAre(t, res, `identifier "n" is not callable (type i64)`)
 }
 
 func TestCall_NonCallable_ResultAssignedToAnnotatedVar_NoError_Cascade(t *testing.T) {
@@ -117,9 +117,9 @@ func TestCall_NonCallable_ResultAssignedToAnnotatedVar_NoError_Cascade(t *testin
 	// bails early with no additional "cannot assign" error.
 	res := parseCollectAndCheck(t, `
 		let n = 42
-		let y: int = n()
+		let y: i64 = n()
 	`, false)
-	assertErrorsAre(t, res, `identifier "n" is not callable (type int)`)
+	assertErrorsAre(t, res, `identifier "n" is not callable (type i64)`)
 }
 
 func TestCall_NonCallable_AnnotatedIntVariable(t *testing.T) {
@@ -135,15 +135,15 @@ func TestCall_NonCallable_AnnotatedIntVariable(t *testing.T) {
 func TestCall_LambdaInVariable_Valid_Ok(t *testing.T) {
 	// Lambda stored in a variable, called with correct args.
 	res := parseCollectAndCheck(t, `
-		let double = (n: int) -> int => n * 2
-		let r: int = double(5)
+		let double = (n: i64) -> i64 => n * 2
+		let r: i64 = double(5)
 	`, false)
 	assertNoErrors(t, res)
 }
 
 func TestCall_LambdaInVariable_WrongArgCount(t *testing.T) {
 	res := parseCollectAndCheck(t, `
-		let double = (n: int) -> int => n * 2
+		let double = (n: i64) -> i64 => n * 2
 		double()
 	`, false)
 	assertErrorsAre(t, res, `double: expected 1 argument(s), got 0`)
@@ -151,37 +151,37 @@ func TestCall_LambdaInVariable_WrongArgCount(t *testing.T) {
 
 func TestCall_LambdaInVariable_WrongArgType(t *testing.T) {
 	res := parseCollectAndCheck(t, `
-		let double = (n: int) -> int => n * 2
-		let r: int = double(true)
+		let double = (n: i64) -> i64 => n * 2
+		let r: i64 = double(true)
 	`, false)
-	assertErrorsAre(t, res, `double: argument 1 (n): cannot assign boolean to int`)
+	assertErrorsAre(t, res, `double: argument 1 (n): cannot assign boolean to i64`)
 }
 
 // ── 4. Direct lambda calls ─────────────────────────────────────────────────
 
 func TestCall_DirectLambda_Valid_Ok(t *testing.T) {
 	// Calling a bare lambda expression directly.
-	res := parseCollectAndCheck(t, `((n: int) -> int => n * 2)(5)`, false)
+	res := parseCollectAndCheck(t, `((n: i64) -> i64 => n * 2)(5)`, false)
 	assertNoErrors(t, res)
 }
 
 func TestCall_DirectLambda_WrongArgCount(t *testing.T) {
-	res := parseCollectAndCheck(t, `((n: int) -> int => n * 2)()`, false)
+	res := parseCollectAndCheck(t, `((n: i64) -> i64 => n * 2)()`, false)
 	assertErrorsAre(t, res, `lambda: expected 1 argument(s), got 0`)
 }
 
 func TestCall_DirectLambda_WrongArgType(t *testing.T) {
-	res := parseCollectAndCheck(t, `((n: int) -> int => n * 2)(true)`, false)
-	assertErrorsAre(t, res, `lambda: argument 1 (n): cannot assign boolean to int`)
+	res := parseCollectAndCheck(t, `((n: i64) -> i64 => n * 2)(true)`, false)
+	assertErrorsAre(t, res, `lambda: argument 1 (n): cannot assign boolean to i64`)
 }
 
 func TestCall_DirectLambda_MultipleParams_Ok(t *testing.T) {
-	res := parseCollectAndCheck(t, `((a: int, b: int) -> int => a + b)(1, 2)`, false)
+	res := parseCollectAndCheck(t, `((a: i64, b: i64) -> i64 => a + b)(1, 2)`, false)
 	assertNoErrors(t, res)
 }
 
 func TestCall_DirectLambda_MultipleParams_WrongCount(t *testing.T) {
-	res := parseCollectAndCheck(t, `((a: int, b: int) -> int => a + b)(1)`, false)
+	res := parseCollectAndCheck(t, `((a: i64, b: i64) -> i64 => a + b)(1)`, false)
 	assertErrorsAre(t, res, `lambda: expected 2 argument(s), got 1`)
 }
 
@@ -189,7 +189,7 @@ func TestCall_DirectLambda_MultipleParams_WrongCount(t *testing.T) {
 
 func TestCall_DefinedFunction_NoArgs_Ok(t *testing.T) {
 	res := parseCollectAndCheck(t, `
-		let answer = () -> int => 42
+		let answer = () -> i64 => 42
 		answer()
 	`, false)
 	assertNoErrors(t, res)
@@ -197,16 +197,16 @@ func TestCall_DefinedFunction_NoArgs_Ok(t *testing.T) {
 
 func TestCall_DefinedFunction_ReturnUsedInAssignment_Ok(t *testing.T) {
 	res := parseCollectAndCheck(t, `
-		let double = (n: int) -> int => n * 2
-		let r: int = double(5)
+		let double = (n: i64) -> i64 => n * 2
+		let r: i64 = double(5)
 	`, false)
 	assertNoErrors(t, res)
 }
 
 func TestCall_DefinedFunction_ChainedCalls_Ok(t *testing.T) {
 	res := parseCollectAndCheck(t, `
-		let inc = (n: int) -> int => n + 1
-		let r: int = inc(inc(0))
+		let inc = (n: i64) -> i64 => n + 1
+		let r: i64 = inc(inc(0))
 	`, false)
 	assertNoErrors(t, res)
 }
@@ -215,8 +215,8 @@ func TestCall_DefinedFunction_UsedBeforeDeclarationOrder_Ok(t *testing.T) {
 	// Both declarations are at the top level; the type-checker processes all
 	// statements in order, so the call site after the declaration is fine.
 	res := parseCollectAndCheck(t, `
-		let square = (n: int) -> int => n * n
-		let r: int = square(4)
+		let square = (n: i64) -> i64 => n * n
+		let r: i64 = square(4)
 	`, false)
 	assertNoErrors(t, res)
 }
